+++ src/utils/index.ts
@@ -28,3 +28,18 @@ export {
   protonHash,
   protonHash64,
 } from './hash.js';
+export {base64Decode, base64Encode} from './base64.js';
+export {
+  binaryDecode,
+  binaryEncode,
+  hexDecode,
+  hexEncode,
+} from './hex-binary.js';
+export {htmlDecode, htmlEncode, urlDecode, urlEncode} from './url-html.js';
+export {
+  MORSE_WORD_SEPARATOR,
+  morseDecode,
+  morseEncode,
+  rot13,
+} from './rot13-morse.js';
+export {decodeJwt} from './jwt.js';